Workig code, fully swiched to new Frame objects
