+++ GUI/AppFrame.java
@@ -1,61 +1,115 @@
-/**
+/*
  * Author: Matěj Šťastný
- * Date created: 4/23/2024
+ * Date created: 
  * Github link: https://github.com/kireiiiiiiii/TargetGame
- * 
- * TODO: Class header
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
  */
 
 package GUI;
 
-import javax.swing.JFrame;
-import java.awt.Dimension;
-import java.awt.Toolkit;
+import javax.swing.*;
 
-import Helpers.*;
-
+/**
+ * A template class of what should be implemented in a custom {@code JFrame}.
+ * TODO: Make class header
+ * 
+ */
 public class AppFrame extends JFrame {
-    private GamePanel panel;
+
+    /////////////////
+    // Fields
+    ////////////////
+
+    private String title;
+    private GamePanel gamePanel;
+
+    /////////////////
+    // Constructors
+    ////////////////
 
     /**
-     * Contructor to be called by {@code SwingUtilities}
+     * Empty contructor. Calls the deafult parameterized constructor with the width
+     * and height zero, and the title of an empty {@code String}.
+     * 
      */
     public AppFrame() {
-        Position windowSize = calculateWindowSize(getScreenDimensions());
-        int windowSizeX = windowSize.getIntX();
-        int windowSizeY = windowSize.getIntY();
-        setSize(windowSizeX, windowSizeY);
+        this(0, 0, "");
+    }
+
+    /**
+     * Deafult parameterized contructor. Initializes the size with given values,
+     * sets the title with given value, sets resizable to {@code false}, sets the
+     * location to the center of the screen and sets the {@code OnCloseOperation} to
+     * {@code EXIT_ON_CLOSE}.
+     * 
+     * @param windowWidth  - width of the {@code JFrame} window.
+     * @param windowHeight - height of the {@code JFrame} window.
+     * @param title        - title of the {@JFrame} window.
+     */
+    public AppFrame(int windowWidth, int windowHeight, String title) {
+        this.title = title;
+        setSize(windowWidth, windowHeight);
         setResizable(false);
-
-        panel = new GamePanel(this);
-        add(panel);
-
-        setTitle("Target Game");
-        setDefaultCloseOperation(EXIT_ON_CLOSE);
+        setTitle(title);
         setLocationRelativeTo(null);
+        setDefaultCloseOperation(EXIT_ON_CLOSE);
         setVisible(true);
+
+        this.gamePanel = new GamePanel(this);
+        add(this.gamePanel);
+
     }
 
-    /**
-     * Calculates the desired size of the window depening on the dimensions of the
-     * user screen.
-     * 
-     * @param screenDimentions - user's screen dimensions.
-     * @return new {@code Position} object
-     */
-    private Position calculateWindowSize(Position screenDimentions) {
-        return screenDimentions.getMultiplied(0.9);
-    }
+    /////////////////
+    // Accesor methods
+    ////////////////
 
     /**
-     * Gets the screen dimensions of the user's screen using {@code Toolkit}
+     * Returns the title of the {@code JFrame} window. 
      * 
-     * @return new {@code Position} object of the dimensions
      */
-    private Position getScreenDimensions() {
-        Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
-        int screenWidth = screenSize.width;
-        int screenHeight = screenSize.height;
-        return new Position(screenWidth, screenHeight);
+    public String getTitle() {
+        return this.title;
     }
+
+    /////////////////
+    // Modifier methods
+    ////////////////
+
+    /**
+     * Changes the title of the {@code JFrame} window. 
+     * 
+     * @param title - new title of the window. 
+     */
+    public void changeTitle(String title) {
+        this.title = title;
+        setTitle(this.title);
+    }
+
+    /////////////////
+    // Other public methods
+    ////////////////
+
+    /////////////////
+    // Private methods
+    ////////////////
 }
